docs: Ignore mock data files
This commit adds the `data/` directory, which contains mock data in JSON format, to the `.gitignore` file.

Ignoring these files prevents them from being committed to the repository, as they are not part of the core application code and are only used for local development. This helps keep the repository clean and focused on essential project files.